feat(week3): Introduce a database helper
Summary
=======

We noticed through the course of the previous run that students got overwhelmed by having to implement both the auth layer as well as dealing with persisting them.

Since the next module is about database, this PR is a proposal to introduce a `database` helper.

## What's in the box

From a student perspective, it adds a bit of reading in the README to understand how to use the database.

It also reshapes the readme to have better sections to help navigate it.

From our standpoint, it introduces lokijs and uuid in the package.json and a database.js file that allow both to create a persistent or in memory database.